requirements.txt
feedparser
typer
django
tqdm
toml
python-dateutil
pydantic
requests
beautifulsoup4
python-dotenv
zulip
